openfind cache with last_modified.

+++ app/controllers/demo/openfind_controller.rb
@@ -13,8 +13,6 @@ class Demo::OpenfindController < Demo::ApplicationController
 
   # Get /demo/openfind
   get "/" do
-    etag md5_key("/demo/openfind static control")
-
     haml :index, layout: settings.layout
   end
 
@@ -22,10 +20,12 @@ class Demo::OpenfindController < Demo::ApplicationController
   post "/members" do
     begin
       filepath = csv_file_for(params[:members][:url])
-      filename = File.basename(filepath)
-      send_file(filepath, :filename => filename)
+      send_file(filepath, filename: File.basename(filepath))
     rescue => e
+      puts e.message
+      puts e.backtrace[0..10]
       flash[:danger] = "请确保链接在浏览器可以正常打开:#{params[:members][:url]}"
+
       redirect "/demo/openfind"
     end
   end
@@ -33,50 +33,56 @@ class Demo::OpenfindController < Demo::ApplicationController
   # Post /demo/openfind/template
   post "/template" do
     begin
-      filepath = dl_template
-      filename = File.basename(filepath)
-      send_file(filepath, :filename => filename)
+      filepath = dl_template(params[:template][:url])
+      send_file(filepath, filename: File.basename(filepath))
     rescue => e
-      flash[:warning] = "请确保链接在浏览器可以正常打开:#{params[:members][:url]}"
+      puts e.message
+      puts e.backtrace[0..10]
+      flash[:danger] = "请确保链接在浏览器可以正常打开:#{params[:template][:url]}"
+
       redirect "/demo/openfind"
     end
   end
 
-  def dl_template
+  def dl_template(url)
+    uri_html = URI.parse(url).read
     tmp_path = File.join(ENV["APP_ROOT_PATH"], "tmp")
     FileUtils.mkdir(tmp_path) if !File.exist?(tmp_path)
-    #创建根目录
-    openfind_path = File.join(tmp_path, "openfind/")
-    File.exist?(openfind_path) ? `rm -fr #{openfind_path}/*` : Dir.mkdir(openfind_path) 
-    images_path = File.join(openfind_path,"images")
-    Dir.mkdir(images_path) unless File.exist?(images_path)
-    
-    #下载image并修改链接地址
-    doc = Nokogiri::HTML(open(params[:template][:url]).read)
-    doc.css("img").each do |img|
-      img_src = img.attr("src")
-      next if !%w(.jpg .jpeg .gif .png).include?(File.extname(img_src).downcase)
-      begin
-        img_src = concat_img_src(params[:template][:url], img_src) if !%r{^(http|https)://(.*)}.match(img_src)
-        img_data = open(img_src){|f| f.read }
-        puts img_src
-      rescue => e
-         puts e.to_s
-      else
-        new_img = File.join(images_path,File.basename(img_src))
-        open(new_img,"wb") { |f| f.write(img_data) }
-        img["src"] = File.join("images",File.basename(img_src))
-      end
-    end
-    #创建index页面
-    File.open(File.join(openfind_path,"index.html"), "wb") { |file| file.puts doc.to_s }
-    zip_file_path = File.join(tmp_path,"openfind #{Time.now.strftime("电子报模板_%Y%m%d%H%M%S")}.zip")
+    last_modified = uri_html.last_modified || Time.now
+    zip_file_path = File.join(tmp_path, last_modified.strftime("OpenfindEDM_%Y%m%d%H%M%S.zip"))
 
-    generate_zip(openfind_path,zip_file_path)
-  #  io = File.open(zf_dir)
-  #  io.binmode
-  #  send_data(io.read,:filename => "openfind #{Time.now.strftime("%Y年%m月电子报模板")}.zip",:disposition => "attachment")
-  #  io.close
+    unless File.exist?(zip_file_path)
+      #创建根目录
+      openfind_path = File.join(tmp_path, "openfind/")
+      File.exist?(openfind_path) ? `rm -fr #{openfind_path}/*` : Dir.mkdir(openfind_path) 
+      images_path = File.join(openfind_path,"images")
+      Dir.mkdir(images_path) unless File.exist?(images_path)
+      
+      #下载image并修改链接地址
+      #doc = Nokogiri::HTML(open(params[:template][:url]).read)
+      doc = Nokogiri::HTML(uri_html)
+      doc.css("img").each do |img|
+        img_src = img.attr("src")
+        next if !%w(.jpg .jpeg .gif .png).include?(File.extname(img_src).downcase)
+        begin
+          img_src = concat_img_src(url, img_src) if !%r{^(http|https)://(.*)}.match(img_src)
+          img_data = open(img_src){|f| f.read }
+          puts img_src
+        rescue => e
+           puts e.to_s
+        else
+          new_img = File.join(images_path,File.basename(img_src))
+          open(new_img,"wb") { |f| f.write(img_data) }
+          img["src"] = File.join("images",File.basename(img_src))
+        end
+      end
+      #创建index页面
+      File.open(File.join(openfind_path,"index.html"), "wb") { |file| file.puts doc.to_s }
+
+      generate_zip(openfind_path,zip_file_path)
+    end
+
+    return zip_file_path
   end
 
   def concat_img_src(url, path)
@@ -95,24 +101,21 @@ class Demo::OpenfindController < Demo::ApplicationController
   end  
   
   def csv_file_for(url)
-    csv_file = Time.now.strftime("Openfind名单_%Y%m%d%H%M%S.csv")
+    uri_html = URI.parse(url).read
     tmp_path = File.join(ENV["APP_ROOT_PATH"], "tmp")
     FileUtils.mkdir(tmp_path) if !File.exist?(tmp_path)
-    csv_path = File.join(tmp_path, csv_file)
+    last_modified = uri_html.last_modified || Time.now
+    csv_path = File.join(tmp_path, last_modified.strftime("OpenfindMembers_%Y%m%d%H%M%S.csv"))
+    puts csv_path
+    unless File.exist?(csv_path)
+      CSV.open(csv_path, "wb:UTF-8") do |csv|
+        csv << ["name", "email"]
+        emails = uri_html.scan(/\w+(?:[\.\-]\w+)*@\w+(?:[\.\-]\w+)*/)
 
-    CSV.open(csv_path, "wb:UTF-8") do |csv|
-      csv << ["name", "email"]
-      emails = open(url).read.scan(/\w+(?:[\.\-]\w+)*@\w+(?:[\.\-]\w+)*/)
-
-      emails.each do |email|
-        csv << [email.split(/@/)[0], email]
-      end if !emails.empty?
-    #  doc = Nokogiri::HTML(open(url).read)
-    #  doc.css("p").each do |p|
-    #   tmp = p.content.strip
-    #   tmp.gsub!(/[" "]|[" "]|[";"]/,"")
-    #   csv << [tmp.split("@")[0], tmp]
-    #  end
+        emails.each do |email|
+          csv << [email.split(/@/)[0], email]
+        end if !emails.empty?
+      end
     end
     return csv_path
   end
